backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
boto3==1.34.101
mistralai==1.0.1
feedparser==6.0.11
python-multipart==0.0.9
aiohttp==3.9.5
rapidfuzz==3.9.3
